+++ TutorialAppKotlin/app/src/main/kotlin/com/generalmagic/gemsdk/demo/activities/mainactivity/controllers/Search.kt
@@ -14,25 +14,20 @@ import android.os.Bundle
 import android.view.View
 import android.widget.SearchView
 import android.widget.Toast
-import com.generalmagic.gemsdk.GenericCategories
-import com.generalmagic.gemsdk.ProgressListener
-import com.generalmagic.gemsdk.SearchService
-import com.generalmagic.gemsdk.demo.activities.LandmarkViewModel
-import com.generalmagic.gemsdk.demo.activities.SLIAdapter
-import com.generalmagic.gemsdk.demo.activities.SearchListActivity
-import com.generalmagic.gemsdk.demo.activities.SearchListItem
+import com.generalmagic.gemsdk.*
+import com.generalmagic.gemsdk.demo.activities.*
+import com.generalmagic.gemsdk.demo.activities.history.Trip
 import com.generalmagic.gemsdk.demo.app.GEMApplication
 import com.generalmagic.gemsdk.demo.app.Tutorials
 import com.generalmagic.gemsdk.demo.util.KeyboardUtil
-import com.generalmagic.gemsdk.models.Coordinates
-import com.generalmagic.gemsdk.models.Landmark
-import com.generalmagic.gemsdk.models.LandmarkCategory
-import com.generalmagic.gemsdk.models.SearchPreferences
+import com.generalmagic.gemsdk.models.*
 import com.generalmagic.gemsdk.util.GEMError
 import com.generalmagic.gemsdk.util.GEMList
 import com.generalmagic.gemsdk.util.GEMSdkCall
 import kotlinx.android.synthetic.main.activity_list_view.*
 import kotlinx.android.synthetic.main.filter_view.*
+import java.util.*
+import kotlin.collections.ArrayList
 
 open class SearchServiceWrapper {
     val service = SearchService()
@@ -75,7 +70,7 @@ open class BaseSearchTutorialActivity : SearchListActivity() {
             val list = GEMSdkCall.execute { results.asArrayList() } ?: ArrayList()
             val reference = GEMSdkCall.execute { preferences.getReferencePoint() }
 
-            val wrapped = wrap(list, reference)
+            val wrapped = wrapLandmarkList(list, reference)
             val adapter = SLIAdapter(wrapped)
 
             list_view.adapter = adapter
@@ -93,23 +88,43 @@ open class BaseSearchTutorialActivity : SearchListActivity() {
         hideProgress()
 
         GEMSdkCall.execute {
-            search.service.cancelRequest(search.listener)
+            search.service.cancelSearch(search.listener)
         }
     }
 
-    private fun wrap(
+    protected fun wrapLandmarkList(
         list: ArrayList<Landmark>, reference: Coordinates?
     ): ArrayList<SearchListItem> {
         val result = ArrayList<SearchListItem>()
         for (item in list) {
             val wrappedItem = LandmarkViewModel(item, reference)
-            wrappedItem.mOnClick = { onItemClicked(wrappedItem) }
+            wrappedItem.mOnClick = { onLandmarkItemClicked(wrappedItem) }
+            wrappedItem.mOnLongClick = {
+                wrappedItem.deleteContent()
+                doStart()
+                true
+            }
             result.add(wrappedItem)
         }
         return result
     }
 
-    protected open fun onItemClicked(item: LandmarkViewModel) {
+    protected fun wrapTripList(list: ArrayList<Trip>): ArrayList<SearchListItem> {
+        val result = ArrayList<SearchListItem>()
+        for ((index, item) in list.withIndex()) {
+            val wrappedItem = TripViewModel(item, index)
+            wrappedItem.mOnClick = { onTripItemClicked(wrappedItem) }
+            wrappedItem.mOnLongClick = { 
+                wrappedItem.deleteContent()
+                doStart()
+                true
+            }
+            result.add(wrappedItem)
+        }
+        return result
+    }
+
+    protected open fun onLandmarkItemClicked(item: LandmarkViewModel) {
         KeyboardUtil.hideKeyboard(this)
         val itLandmark = item.it ?: return
 
@@ -117,6 +132,33 @@ open class BaseSearchTutorialActivity : SearchListActivity() {
 
         /*DEFAULT*/
         Tutorials.openWikiTutorial(itLandmark)
+    }
+
+    protected open fun onTripItemClicked(item: TripViewModel) {
+        KeyboardUtil.hideKeyboard(this)
+        val itTrip = item.it ?: return
+
+        finish()
+        itTrip.mWaypoints?.let { waypoints ->
+
+            if (!itTrip.mIsFromAToB) {
+                val list = arrayListOf<Landmark>()
+                val position = PositionService().getPosition()
+                if (position != null) {
+                    list.add(
+                        Landmark(
+                            "",
+                            Coordinates(position.getLatitude(), position.getLongitude())
+                        )
+                    )
+                }
+
+                list.addAll(waypoints)
+                Tutorials.openCustomRouteTutorial(list)
+                return@let
+            }
+            Tutorials.openCustomRouteTutorial(waypoints)
+        }
     }
 
     protected fun getCurrentCoords(): Coordinates? = GEMSdkCall.execute {
@@ -188,6 +230,110 @@ class SearchNearbyActivity : BaseSearchTutorialActivity() {
                 search.results, search.listener, coords, "", preferences
             )
         }
+    }
+}
+
+class SearchHistoryActivity : BaseSearchTutorialActivity() {
+    private val listToDisplay = arrayListOf<SearchListItem>()
+    private val searchViewListener = object : SearchView.OnQueryTextListener {
+        override fun onQueryTextSubmit(text: String?): Boolean {
+            if (text != null) {
+                doSearch(text)
+            }
+            searchInput?.clearFocus()
+            return true
+        }
+
+        override fun onQueryTextChange(text: String?): Boolean {
+            if (text != null) {
+                doSearch(text)
+            }
+            return true
+        }
+    }
+    
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        filterView.visibility = View.VISIBLE
+        searchInput.requestFocus()
+        searchInput.setOnQueryTextListener(searchViewListener)
+    }
+
+    override fun doStart() {
+        doStop()
+
+        GEMSdkCall.execute {
+            val coords = getCurrentCoords() ?: return@execute
+
+            preferences.setReferencePoint(coords)
+            preferences.setSearchMapPOIs(true)
+
+//            val listToDisplay = arrayListOf<SearchListItem>()
+            GEMApplication.getHistoryLandmarkStore()?.getLandmarks()?.let { list ->
+                val reference = GEMSdkCall.execute { preferences.getReferencePoint() }
+                val wrapped = wrapLandmarkList(list, reference)
+                listToDisplay.addAll(wrapped)
+            }
+
+            GEMApplication.getTripsHistory()?.let { tripsHistory ->
+                val nItems = tripsHistory.getTripsCount()
+                val list = arrayListOf<Trip>()
+                for (index in 0 until nItems) {
+                    val trip = tripsHistory.loadTrip(index).second
+                    if (trip != null) {
+                        list.add(trip)
+                    }
+                }
+                val wrapped = wrapTripList(list)
+                listToDisplay.addAll(wrapped)
+            }
+
+            listToDisplay.sortByDescending { it.getId() }
+            val adapter = HLIAdapter(listToDisplay)
+
+            GEMApplication.postOnMain { list_view.adapter = adapter }
+
+        }
+    }
+    
+    private fun doSearch(filter: String) {
+        val resultList = if (filter.isNotEmpty()) {
+            val lowerFilter = filter.toLowerCase(Locale.getDefault())
+            val filterTokens = lowerFilter.split(' ', '-')
+
+            ArrayList(
+                listToDisplay.filter {
+                    val arg = GEMSdkCall.execute { it.getText() }
+                    val lowerArg = arg?.toLowerCase(Locale.getDefault()) ?: ""
+                    val argTokens = lowerArg.split(' ', '-')
+
+                    for (filterWord in filterTokens) {
+                        var found = false
+                        for (token in argTokens) {
+                            if (token.contains(filterWord)) {
+                                found = true
+                                break
+                            }
+                        }
+                        if (!found) {
+                            return@filter false
+                        }
+                    }
+
+                    true
+                }
+            )
+        } else {
+            listToDisplay
+        }
+
+        displayList(resultList)
+    }
+
+    fun displayList(items: ArrayList<SearchListItem>) {
+        val adapter = HLIAdapter(items)
+
+        list_view.adapter = adapter
     }
 }