+++ TutorialAppKotlin/app/src/main/kotlin/com/generalmagic/gemsdk/demo/activities/searchaddress/GEMAddressSearchView.kt
@@ -148,7 +148,7 @@ object GEMAddressSearchView {
             mField = field
 
             mListener?.let {
-                GEMSdkCall.execute { GuidedAddressSearchService().cancelRequest(it) }
+                GEMSdkCall.execute { GuidedAddressSearchService().cancelSearch(it) }
             }
 
             search(getParentLandmark(field), filter, getDetailLevel(field))